fix decoding problem with wide instructions

+++ src/main/java/net/oneandone/mork/classfile/InstructionEncoding.java
@@ -115,7 +115,7 @@ public class InstructionEncoding implements Bytecodes, Constants {
         switch (b) {
         case IINC:
             args = new Object[2];
-            args[0] = src.readU2();
+            args[0] = (int) src.readU2();
             args[1] = src.readS2();
             break;
         case ILOAD:
@@ -130,7 +130,7 @@ public class InstructionEncoding implements Bytecodes, Constants {
         case DSTORE:
         case RET:
             // TODO: don't hard-wire constants
-            args = new Object[] { src.readU2() };
+            args = new Object[] { (int) src.readU2() };
             break;
         default:
             throw new RuntimeException("illegal iinc");
